IA Monster + Deplacement

+++ RPG/src/playable/MoveTreatment.java
@@ -10,22 +10,22 @@ public class MoveTreatment {
 		case 0:
 			ply.setDirection(direction);
 			if (ply.getY() - mvt >= 0) {
-				if (map.isOccupied(ply.getX(), ply.getY() - mvt)) {
+				if (map.isOccupied((int)ply.getX(), (int)ply.getY() - mvt)) {
 					System.out.println("Case occupée ! C'est le monstre :"
-							+ map.getMonsterNameByPos(ply.getX(), ply.getY() - mvt));
+							+ map.getMonsterNameByPos((int)ply.getX(),(int) ply.getY() - mvt));
 				}
-				else if (map.isProp(ply.getX(), ply.getY() - mvt)) {
+				else if (map.isProp((int)ply.getX(),(int) ply.getY() - mvt)) {
 					System.out.println("Case occupée par un prop");
 				}
 				else {
 					ply.setY(ply.getY() - mvt);
-					System.out.println("X : " + ply.getX() + " / Y : " + ply.getY() + " / Regard : "
+					System.out.println("X : " + (int)ply.getX() + " / Y : " + (int)ply.getY() + " / Regard : "
 							+ charDirection(direction) + " - " + ply.getDirection());
 				}
 			} else {
 				System.out.println("BLOQUÉ PAR UN MUR");
 				ply.setY(0);
-				System.out.println("X : " + ply.getX() + " / Y : " + ply.getY() + " / Regard : "
+				System.out.println("X : " +(int) ply.getX() + " / Y : " +(int) ply.getY() + " / Regard : "
 						+ charDirection(direction) + " - " + ply.getDirection());
 			}
 			break;
@@ -33,11 +33,11 @@ public class MoveTreatment {
 		case 3:
 			ply.setDirection(direction);
 			if (ply.getY() + mvt <= map.getWidth()) {
-				if (map.isOccupied(ply.getX(), ply.getY() + mvt)) {
+				if (map.isOccupied((int)ply.getX(),(int) ply.getY() + mvt)) {
 					System.out.println("Case occupée ! C'est le monstre :"
-							+ map.getMonsterNameByPos(ply.getX(), ply.getY() + mvt));
+							+ map.getMonsterNameByPos((int)ply.getX(),(int) ply.getY() + mvt));
 				} 
-				else if (map.isProp(ply.getX(), ply.getY() + mvt)) {
+				else if (map.isProp((int)ply.getX(),(int) ply.getY() + mvt)) {
 					System.out.println("Case occupée par un prop");
 				}
 				else {
@@ -57,11 +57,11 @@ public class MoveTreatment {
 			ply.setDirection(direction);
 			if (ply.getX() - mvt >= 0) {
 				System.out.println("hi");
-				if (map.isOccupied(ply.getX() - mvt, ply.getY())) {
+				if (map.isOccupied((int)ply.getX() - mvt, (int)ply.getY())) {
 					System.out.println("Case occupée ! C'est le monstre :"
-							+ map.getMonsterNameByPos(ply.getX() - mvt, ply.getY()));
+							+ map.getMonsterNameByPos((int)ply.getX() - mvt, (int)ply.getY()));
 				} 
-				else if (map.isProp(ply.getX() - mvt, ply.getY())) {
+				else if (map.isProp((int)ply.getX() - mvt, (int)ply.getY())) {
 					System.out.println("Case occupée c'est un prop");
 				}
 				else {
@@ -80,11 +80,11 @@ public class MoveTreatment {
 		case 2:
 			ply.setDirection(direction);
 			if (ply.getX() + mvt <= map.getLength()) {
-				if (map.isOccupied(ply.getX() + mvt, ply.getY())) {
+				if (map.isOccupied((int)ply.getX() + mvt, (int)ply.getY())) {
 					System.out.println("Case occupée ! C'est le monstre : "
-							+ map.getMonsterNameByPos(ply.getX() + mvt, ply.getY()));
+							+ map.getMonsterNameByPos((int)ply.getX() + mvt,(int) ply.getY()));
 				}
-				else if (map.isProp(ply.getX() + mvt, ply.getY())) { 
+				else if (map.isProp((int)ply.getX() + mvt,(int) ply.getY())) { 
 					System.out.println("Case occupée c'est un prop");
 				}
 				else {
